% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGoogleMap}
\alias{getGoogleMap}
\title{Get Google Man using ggmap function}
\usage{
getGoogleMap(lon, lat, zoom, prefix = "google")
}
\arguments{
\item{lon}{Longitude}

\item{lat}{Latitude}

\item{zoom}{Zoom}

\item{prefix}{Prefix for the data saved}
}
\description{
Get Google Man using ggmap function
}
\details{
This fucntion will get the Google map from the
}
\author{
Bingwei Tian
}

